requirements.txt
aioboto3==12.3.0
aiobotocore==2.11.2
aiohttp==3.9.1
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.37.1
anyio==3.7.1
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
blinker==1.9.0
boto3==1.34.34
botocore==1.34.34
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
comm==0.2.2
cryptography==41.0.7
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.109.0
filelock==3.17.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2025.2.0
functions-framework==3.8.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-secret-manager==2.16.1
googleapis-common-protos==1.67.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
openai==1.3.3
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.19.1
pyOpenSSL==23.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.35.1
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.0.7
uvicorn==0.27.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
